% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phinterval.R
\name{is_phinterval}
\alias{is_phinterval}
\title{Test if the object is a phinterval}
\usage{
is_phinterval(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is a \verb{<phinterval>}, \code{FALSE} otherwise.
}
\description{
This function returns \code{TRUE} for \verb{<phinterval>} vectors and returns
\code{FALSE} otherwise.
}
\examples{
is_phinterval(phinterval())
is_phinterval(interval())

}
