% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phinterval.R
\name{is_phintish}
\alias{is_phintish}
\title{Test if the object is a phinterval or interval}
\usage{
is_phintish(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is a \verb{<phinterval>} or \verb{<Interval>}, \code{FALSE} otherwise.
}
\description{
This function returns \code{TRUE} for \verb{<phinterval>} and \verb{<Interval>} vectors and
returns \code{FALSE} otherwise.
}
\examples{
is_phinterval(phinterval())
is_phinterval(interval())
is_phinterval(as.Date("2020-01-01"))

}
