% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-modifications.R
\name{phint_discard_instants}
\alias{phint_discard_instants}
\title{Remove instantaneous time spans from a phinterval}
\usage{
phint_discard_instants(phint)
}
\arguments{
\item{phint}{\verb{[phinterval / Interval]}

A \verb{<phinterval>} or \verb{<Interval>} vector.}
}
\value{
A \verb{<phinterval>} vector the same length as \code{phint}.
}
\description{
\code{phint_discard_instants()} removes instantaneous spans (spans with 0 duration) from
phinterval elements. If all spans in an element are instantaneous, the result
is a hole.
}
\examples{
y2020 <- interval(as.Date("2020-01-01"), as.Date("2021-01-01"))
y2021 <- interval(as.Date("2021-01-01"), as.Date("2022-01-01"))
y2022 <- interval(as.Date("2022-01-01"), as.Date("2023-01-01"))

# The intersection of two adjacent intervals is instantaneous
new_years_2021 <- phint_intersect(y2020, y2021)
new_years_2021
phint_discard_instants(new_years_2021)

# phint_discard_instants() removes instants while keeping non-instantaneous spans
y2022_and_new_years <- phint_union(y2022, new_years_2021)
y2022_and_new_years
phint_discard_instants(y2022_and_new_years)

}
