% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phinterval.R
\name{phint_invert}
\alias{phint_invert}
\alias{phint_invert.default}
\alias{phint_invert.Interval}
\alias{phint_invert.phinterval}
\title{Get the gaps in a phinterval as time spans}
\usage{
phint_invert(phint, hole_to = c("hole", "inf", "na"))

\method{phint_invert}{default}(phint, hole_to = c("hole", "inf", "na"))

\method{phint_invert}{Interval}(phint, hole_to = c("hole", "inf", "na"))

\method{phint_invert}{phinterval}(phint, hole_to = c("hole", "inf", "na"))
}
\arguments{
\item{phint}{\verb{[phinterval / Interval]}

A \verb{<phinterval>} or \verb{<Interval>} vector.}

\item{hole_to}{\verb{["hole" / "inf" / "na"]}

What to turn \verb{<hole>} (i.e. empty) time spans into.
\itemize{
\item If \code{hole_to = "hole"} (the default), \verb{<hole>} spans remain as \verb{<hole>} elements.
\item If \code{"inf"}, they are returned as a time span from \code{-Inf} to \code{Inf}.
\item If \code{"na"}, they are returned as a missing (\code{NA}) span.
}}
}
\value{
A \verb{<phinterval>} vector the same length as \code{phint}.
}
\description{
\code{phint_invert()} returns the gaps in a phinterval as a \verb{<phinterval>} vector.
Contiguous time spans (e.g. \code{\link[lubridate:interval]{lubridate::interval()}} vectors) are inverted to
\verb{<hole>} time spans.

\code{phint_invert()} is similar to \code{phint_complement()}, except that the time occurring
outside the extent of \code{phint} (i.e. before its earliest start or after its
latest end) is not included in the result.
}
\examples{
monday <- interval(as.Date("2025-11-10"), as.Date("2025-11-11"))
friday <- interval(as.Date("2025-11-14"), as.Date("2025-11-15"))
sunday <- interval(as.Date("2025-11-16"), as.Date("2025-11-17"))

# Contiguous intervals are inverted to holes,
# disjoint intervals to spans
phint_invert(monday)
phint_invert(phint_squash(c(monday, friday, sunday))

tues_to_thurs <- interval(as.Date("2025-11-11"), as.Date("2025-11-14"))
phint_invert(phint_union(monday, friday)) == tues_to_thurs

# The time before `monday` and after `friday` is included
# in the complement, but not the inversion
mon_and_fri <- phint_union(monday, friday)
phint_invert(mon_and_fri)
phint_complement(mon_and_fri)

# Specify how to invert empty time spans
hole <- phint_intersect(monday, friday)
phint_invert(hole, hole_to = "hole")
phint_invert(hole, hole_to = "inf")
phint_invert(hole, hole_to = "na")

}
