% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-operations.R
\name{phint_squash}
\alias{phint_squash}
\title{Flatten a phinterval}
\usage{
phint_squash(phint, na.rm = TRUE, empty_to = c("na", "hole", "empty"))
}
\arguments{
\item{phint}{\verb{[phinterval / Interval]}

A \verb{<phinterval>} or \verb{<Interval>} vector.}

\item{na.rm}{\verb{[TRUE / FALSE]}

Whether to ignore \code{NA} elements when flattening a \verb{<phinterval>}.

If \code{na.rm = FALSE} and \code{phint} contains any \code{NA} elements, then a scalar \code{NA}
\verb{<phinterval>} will be returned.}

\item{empty_to}{What to return if the input \code{phint} is length 0.
\itemize{
\item If \code{empty_to = "na"} (the default), a length 1 \code{NA} \verb{<phinterval>} is returned.
\item If \code{"hole"} a single \verb{<hole>} is returned.
\item If \code{"empty"} a length 0 \verb{<phinterval>} vector is returned.
}}
}
\value{
A length 1 \verb{<phinterval>} vector, with one exception.

If the input is an empty \verb{<phinterval>} or \verb{<Interval>} vector \emph{and}
\code{empty_to = "empty"}, then the result is a length 0 \verb{<phinterval>} vector.
}
\description{
\code{phint_squash()} combines all elements of a \verb{<phinterval>} vector into a
single set of non-overlapping, non-adjacent time spans, returned as a
scalar \verb{<phinterval>} by default.

This is useful in conjunction with aggregation functions, e.g.
\code{\link[dplyr:summarise]{dplyr::summarize()}}).
}
\examples{
jan_1_to_5   <- interval(as.Date("2000-01-01"), as.Date("2000-01-05"))
jan_3_to_9   <- interval(as.Date("2000-01-03"), as.Date("2000-01-09"))
jan_11_to_12 <- interval(as.Date("2000-01-11"), as.Date("2000-01-12"))

phint_squash(c(jan_1_to_5, jan_3_to_9, jan_11_to_12, NA))

# Set `na.rm = FALSE` to propagate `NA` values
phint_squash(c(jan_1_to_5, jan_3_to_9, jan_11_to_12, NA), na.rm = FALSE)

# `empty_to` determines the result of empty inputs
empty <- interval()
phint_squash(empty, empty_to = "na")
phint_squash(empty, empty_to = "hole")
phint_squash(empty, empty_to = "empty")

}
