% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phinterval.R
\name{phint_to_spans}
\alias{phint_to_spans}
\title{Convert a phinterval into a list of intervals}
\usage{
phint_to_spans(phint, hole_to = c("empty", "na", "null"))
}
\arguments{
\item{phint}{\verb{[phinterval / Interval]}

A \verb{<phinterval>} or \verb{<Interval>} vector.}

\item{hole_to}{\verb{["empty" / "na" / "null"]}

What to turn \verb{<hole>} (i.e. empty) time spans into.
If \code{hole_to = "empty"} (the default), \verb{<hole>} spans are returned as length 0 intervals.
If \code{"na"}, they are returned as a missing (\code{NA}) interval.
If \code{"null"}, they are returned as a \code{NULL} element.}
}
\value{
A list of \verb{<Interval>} vectors the same length as \code{phint}. If \code{hole_to = "null"},
the list may contain \code{NULL} elements.
}
\description{
\code{phint_to_spans()} decomposes each element of a phinterval into the set
of its contiguous time spans, returned as \code{\link[lubridate:interval]{lubridate::interval()}} vectors.
}
\examples{
monday <- interval(as.Date("2025-11-10"), as.Date("2025-11-11"))
friday <- interval(as.Date("2025-11-14"), as.Date("2025-11-15"))
sunday <- interval(as.Date("2025-11-16"), as.Date("2025-11-17"))

mon_and_fri <- phint_union(monday, friday)
phint_to_spans(mon_and_fri)
phint_to_spans(sunday)

# Specify how to invert empty time spans
hole <- phint_intersect(monday, friday)
phint_to_spans(hole, hole_to = "empty")
phint_to_spans(hole, hole_to = "na")
phint_to_spans(hole, hole_to = "null")

}
