% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phinterval.R
\name{phinterval-span-predicates}
\alias{phinterval-span-predicates}
\alias{is_hole}
\alias{is_span}
\alias{is_disjoint}
\title{Test for empty, contiguous, or disjoint intervals}
\usage{
is_hole(phint)

is_span(phint)

is_disjoint(phint)
}
\arguments{
\item{phint}{\verb{[phinterval / Interval]}

A \verb{<phinterval>} or \verb{<Interval>} vector.}
}
\value{
A logical vector the same length as \code{phint}.
}
\description{
These predicates test structural properties of each element of a
\verb{<phinterval>} vector.
\itemize{
\item \code{is_hole()}: Is the element empty (zero spans)?
\item \code{is_span()}: Is the element a single contiguous span?
\item \code{is_disjoint()}: Is the element made up of two or more disjoint spans?
}
}
\section{Under Development}{


\code{is_span()} and \code{is_disjoint()} are under development and may contain bugs. To
use these functions, install the development version of {phinterval} from GitHub
with \code{pak::pak("EthanSansom/phinterval")}.
}

\examples{
y2000 <- interval(as.Date("2000-01-01"), as.Date("2001-01-01"))
y2025 <- interval(as.Date("2025-01-01"), as.Date("2026-01-01"))
holey <- phint_union(y2000, y2025)

# Detect holes
is_hole(c(hole(), y2000, NA))

# The intersection of disjoint intervals is a hole
is_hole(phint_intersect(y2000, y2025))

# Detect single contiguous spans
is_span(c(y2000, holey, hole(), NA))

# Detect disjoint (multi-span) elements
is_disjoint(c(y2000, holey, hole(), NA))

}
