% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phinterval.R
\name{phinterval}
\alias{phinterval}
\title{Create a new phinterval}
\usage{
phinterval(intervals = list(), tzone = NULL)
}
\arguments{
\item{intervals}{\verb{[list of Interval]}

A list of \code{\link[lubridate:interval]{lubridate::interval()}} vectors. Each interval vector is merged
into a set of non-overlapping and non-adjacent time spans. Empty (length 0)
interval vectors are returned as an empty set of time spans (i.e. a \verb{<hole>}).}

\item{tzone}{\verb{[character(1)]}

A recognized timezone to display the \verb{<phinterval>} in. If \code{tzone} is \code{NULL}
(the default), then the timezone of the first element of \code{intervals} is used.}
}
\value{
A \verb{<phinterval>} vector.
}
\description{
\code{phinterval()} creates a new \verb{<phinterval>} vector from a list of
\code{\link[lubridate:interval]{lubridate::interval()}} vectors. A phinterval (think "potentially holey interval")
is a span of time which may contain gaps.
}
\details{
The \verb{<phinterval>} class is designed as a generalization of the
\link[lubridate:Interval-class]{lubridate::Interval}. While an \verb{<Interval>} element represents a
single contiguous span between two fixed times, a \verb{<phinterval>} element can
represent a time span that may be empty, contiguous, or disjoint (i.e. containing
gaps). Each element of a \verb{<phinterval>} is stored as a (possibly empty) set of
non-overlapping and non-abutting time spans.
}
\examples{
jan <- interval(as.Date("2000-01-01"), as.Date("2000-02-01"), tz = "UTC")
feb <- interval(as.Date("2000-02-01"), as.Date("2000-03-01"), tz = "UTC")
nov <- interval(as.Date("2000-11-01"), as.Date("2000-12-01"), tz = "UTC")

phinterval(list(jan, c(feb, nov)))
phinterval(list(jan), tzone = "EST")
phinterval()

# Empty (length 0) intervals can be used to create <hole> elements
phinterval(list(interval()))

# Abutting or overlapping intervals are merged into one span
phinterval(list(c(jan, feb)))

}
