% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phinterval.R
\name{\%within\%}
\alias{\%within\%}
\alias{\%within\%,phinterval,phinterval-method}
\alias{\%within\%,ANY,phinterval-method}
\alias{\%within\%,Interval,phinterval-method}
\alias{\%within\%,phinterval,Interval-method}
\title{Does a date or phinterval fall within a phinterval}
\usage{
a \%within\% b
}
\arguments{
\item{a}{\verb{[phinterval / Interval / instant]}

A phinterval, interval, or date-time object.}

\item{b}{\verb{[phinterval / Interval]}

A phinterval or interval vector.}
}
\value{
A logical vector the same length as \code{a}.
}
\description{
Check whether \code{a} is within the phinterval \code{b}, inclusive of
endpoints. If \code{a} is a phinterval, then all of it's spans must lie within
the span(s) of \code{b}.
}
\examples{
library(lubridate)
scheduled_break <- interval(ymd_hm("2020-01-01 12:00"), ymd_hm("2020-01-01 12:30"))
scheduled_shift_end <- ymd_hm("2020-01-01 17:00")

morning_worked <- interval(ymd_hm("2020-01-01 09:00"), ymd_hm("2020-01-01 13:00"))
afternoon_worked <- interval(ymd_hm("2020-01-01 12:31"), ymd_h("2020-01-01 16:30"))
shift_worked <- phinterval(list(morning_worked, afternoon_worked))

}
