% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phinterval.R
\name{as_phinterval}
\alias{as_phinterval}
\alias{as_phinterval.default}
\alias{as_phinterval.Interval}
\title{Convert an interval vector into a phinterval}
\usage{
as_phinterval(x, ...)

\method{as_phinterval}{default}(x, ...)

\method{as_phinterval}{Interval}(x, ...)
}
\arguments{
\item{x}{An object to convert.}

\item{...}{Parameters passed to other methods. Currently unused.}
}
\value{
A \verb{<phinterval>} vector the same length as \code{x}.
}
\description{
\code{as_phinterval()} changes a \code{\link[lubridate:interval]{lubridate::interval()}} vector into the equivalent
\verb{<phinterval>} vector. Negative intervals are flipped to positive (i.e.
via \code{\link[lubridate:interval]{lubridate::int_standardize()}}).
}
\examples{
years <- interval(
  start = as.Date(c("2021-01-01", "2023-01-01")),
  end =  as.Date(c("2022-01-01", "2024-01-01"))
)
as_phinterval(years)

# Negative intervals are standardized
(negative <- interval(as.Date("2000-10-11"), as.Date("2000-10-11") - 10))
as_phinterval(negative)

# A <phinterval> cannot have partially missing endpoints
(partial_na <- interval(NA, as.Date("1999-08-02")))
as_phinterval(partial_na)

}
\seealso{
\code{\link[=phinterval]{phinterval()}}
}
