% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\alias{parameters}
\alias{arguments}
\alias{args}
\title{Parameter Descriptions for phinterval Package}
\arguments{
\item{phint1, phint2}{\verb{[phinterval / Interval]}

A pair of \verb{<phinterval>} or \verb{<Interval>} vectors of the same length or of
length 1. Inputs of length 1 are recycled.}

\item{phint}{\verb{[phinterval / Interval]}

A \verb{<phinterval>} or \verb{<Interval>} vector.}
}
\value{
The value \code{NULL}.
}
\description{
Default parameter descriptions which may be overridden in individual
functions.
}
\keyword{internal}
