% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-relations.R
\name{phint_overlaps}
\alias{phint_overlaps}
\title{Test whether two phintervals overlap}
\usage{
phint_overlaps(phint1, phint2)
}
\arguments{
\item{phint1, phint2}{\verb{[phinterval / Interval]}

A pair of \verb{<phinterval>} or \verb{<Interval>} vectors of the same length or of
length 1. Inputs of length 1 are recycled.}
}
\value{
A logical vector.
}
\description{
\code{phint_overlaps()} tests whether the i-th element of \code{phint1} overlaps
with the i-th element of \code{phint2}, returning a logical vector. Adjacent
intervals are considered overlapping. \code{phint1} and \code{phint2} are recycled to
their common length using vctrs-style recycling rules.
}
\examples{
monday <- interval(as.Date("2025-11-10"), as.Date("2025-11-11"))
tuesday <- interval(as.Date("2025-11-11"), as.Date("2025-11-12"))
friday <- interval(as.Date("2025-11-14"), as.Date("2025-11-15"))
mon_and_fri <- phint_union(monday, friday)

phint_overlaps(c(monday, monday, friday), c(mon_and_fri, friday, NA))

# Adjacent intervals are considered overlapping
phint_overlaps(monday, tuesday)

# Holes are always considered non-overlapping
hole <- phinterval(interval())
phint_overlaps(c(hole, monday), c(hole, hole))

}
