% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-modifications.R
\name{phint_sift}
\alias{phint_sift}
\title{Remove instantaneous time spans from a phinterval}
\usage{
phint_sift(phint)
}
\arguments{
\item{phint}{\verb{[phinterval / Interval]}

A \verb{<phinterval>} or \verb{<Interval>} vector.}
}
\value{
A \verb{<phinterval>} vector the same length as \code{phint}.
}
\description{
\code{phint_sift()} removes instants (i.e. spans of 0 seconds in duration) from
elements of a phinterval.
}
\examples{
y2020 <- interval(as.Date("2020-01-01"), as.Date("2021-01-01"))
y2021 <- interval(as.Date("2021-01-01"), as.Date("2022-01-01"))
y2022 <- interval(as.Date("2022-01-01"), as.Date("2023-01-01"))

# The intersection of two adjacent intervals is an instant.
# phint_sift() is useful for removing these instants.
(new_years_2021 <- phint_intersect(y2020, y2021))
phint_sift(new_years_2021)

(y2022_and_new_years_2021 <- phint_union(y2022, new_years_2021))
phint_sift(y2022_and_new_years_2021)

}
