% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-operations.R
\name{phint_within}
\alias{phint_within}
\title{Test whether a date, time, or phinterval is within another phinterval}
\usage{
phint_within(x, phint)
}
\arguments{
\item{x}{A \verb{<POSIXct>}, \verb{<POSIXlt>}, \verb{<Date>}, \verb{<Interval>} or \verb{<phinterval>}
vector to test.}

\item{phint}{\verb{[phinterval / Interval]}

A \verb{<phinterval>} or \verb{<Interval>} vector.}
}
\value{
A logical vector.
}
\description{
\code{phint_within()} tests whether the i-th element of \code{x} is contained within
the i-th element of \code{phint}, returning a logical vector. \code{x} may be a date or
time, while \code{phint} must be a phinterval. \code{x} and \code{phint} are recycled to
their common length using vctrs-style recycling rules.

Dates and times on an endpoint of an interval are considered to be within
the interval. An interval is considered to be within itself.
}
\examples{
jan_1_to_5 <- interval(as.Date("2000-01-01"), as.Date("2000-01-05"))
jan_2_to_4 <- interval(as.Date("2000-01-03"), as.Date("2000-01-09"))
jan_3_to_9 <- interval(as.Date("2000-01-03"), as.Date("2000-01-09"))

phint_within(
  c(jan_2_to_4, jan_3_to_9, jan_1_to_5),
  c(jan_1_to_5, jan_1_to_5, NA)
)

phint_within(as.Date(c("2000-01-06", "2000-01-20")), jan_3_to_9)

# Intervals are within themselves
phint_within(jan_1_to_5, jan_1_to_5)

# Interval endpoints are considered to be within the interval
phint_within(as.Date("2000-01-01"), jan_1_to_5)

# Holes are never considered to be within an interval
hole <- phinterval(interval())
phint_within(c(hole, hole), c(hole, jan_1_to_5))

}
