% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phinterval-package.R
\docType{package}
\name{phinterval-package}
\alias{phinterval}
\alias{phinterval-package}
\title{phinterval: Set Operations for Time Intervals}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implements the phinterval vector class for representing time spans that may contain gaps (disjoint intervals) or be empty. This class generalizes the 'lubridate' package's interval class to support vectorized set operations (intersection, union, difference, complement) that always return a valid time span, even when gaps or empty sets are created.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EthanSansom/phinterval}
  \item \url{https://ethansansom.github.io/phinterval/}
  \item Report bugs at \url{https://github.com/EthanSansom/phinterval/issues}
}

}
\author{
\strong{Maintainer}: Ethan Sansom \email{ethan.sansom29@gmail.com} (\href{https://orcid.org/0009-0000-1573-0186}{ORCID}) [copyright holder]

}
\keyword{internal}
